==> 0-importacion.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

# Importacion {#import}


```{r message=FALSE, warning=FALSE, eval=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
pacman::p_load(tidyverse, googlesheets4, googledrive, lubridate)
gs4_auth(email = "quipu.dat@gmail.com")
googledrive::drive_auth(email = "quipu.dat@gmail.com")
```

## Soja

Conectar con google sheet

```{r}
sheet_2017_2021_cba_soja <- gs4_get(gs4_find("2017-2021_cba_soja")$id)
# sheet_2017_2021_cba_soja %>% gs4_browse()
```

Importar hoja de datos (raw)

```{r}
raw <- read_sheet(sheet_2017_2021_cba_soja, 
                  sheet = "raw", guess_max = 10000, skip=0) %>% 
  janitor::clean_names() 
```

```{r}
raw %>% glimpse
```

# raw to dat

codificar a la nomenclatura de los scripts

```{r, eval=FALSE}
raw %>% 
  mutate_if(is.character, 
           list(
             #sacar acentos
             ~stringi::stri_trans_general(
             #pasar a mayusculas y sacar puntos
             str_to_upper(gsub(',', '\\.', 
             # sacar espacios antes-desp
             str_trim(str_replace_all(., fixed(" "), "_")))), "Latin-ASCII"))
    )%>% 
 rename(
   #aca debemos convertir los nombres de las col para nuestro coding
   geno_raw = variedad
   ) %>% 
  # mutate(Zona = case_when(
  #   str_detect(zona_0, "Madariaga|Maipú") ~ "Madariaga", 
  #   str_detect(zona_0, "Balcarce") ~ "Sierra", 
  #   TRUE ~ "Costa"
  # )) %>%  
  #  mutate(
  #   year = factor(lubridate::year(fecha_de_siembra_dia_mes)),     
  #   date = update(fecha_de_siembra_dia_mes, year = 1), 
  #   temprano_tardio= case_when(
  #   date<as.Date("0001-07-15") ~ "Temprano", 
  #   date>as.Date("0001-07-15") ~ "Tardio")) %>% 
  # rowwise() %>%
  # mutate(dano_tot = sum(dano_1,dano_2, na.rm = T)) %>% 
 # mutate_at(vars(
 #   localidad, variedad,calidad_del_lote, cultivo_de_cosecha, antecesor_estival,fungicida1, fungicida2, enfermedad_1, enfermedad_2), as.factor) %>% 
ungroup -> dat
```

Generar un script exclusivo de genotipos

```{r, eval=FALSE}
geno1 <- read_sheet(sheet_2017_2021_cba_soja, sheet = "geno_raw") %>% 
  # drop_na(geno_new, gm_new) %>% 
  select(geno_raw, geno_new)

dat_geno  <- dat %>% 
  filter(!geno_raw %in% (geno1 %>% filter(geno_new=="borrar") %>% pull(geno_raw)))  %>% 
  drop_na(geno_raw) %>%  
  drop_na(gm) 

dat_geno <- geno1 %>% 
  drop_na(geno_new) %>%
  filter(!geno_new=="borrar") %>%
  full_join(dat_geno, by = c("geno_raw")) %>%
  mutate(geno=coalesce(geno_new, geno_raw)) %>% 
  mutate(geno= str_replace(geno, "_IPRO|_IRPO|_STS|STS", ""))
```

Crear una hoja para mostrar como fue depurado

```{r}
dat_geno %>% 
  count(gm, geno)%>% 
  write_sheet(ss=sheet_2017_2021_cba_soja, sheet = "geno_clean")
```

```{r}
cs <- read_sheet(sheet_2017_2021_cba_soja, 
                  sheet = "CS", guess_max = 10000, skip=0) %>% 
  janitor::clean_names() 
```

```{r, eval=FALSE}
load(here::here("data/data.Rdata"))
save(raw, dat, dat_geno, cs, file = "data/data.Rdata")
```

```{r eval = F}
glimpse(dat)
```

```{r eval = F}
all_images <- list.files(path = here::here("plots"), 
                         pattern = ".png", 
                         all.files = TRUE, full.names = TRUE)
all_images %>% 
  map(~ drive_upload(.,
                     path = as_dribble("juanchi_guille/JMF_fina_2020"),
                     overwrite = TRUE)
  )
```

```{r}
pacman::p_load(tidyverse, lubridate, zoo)

serie_tandil0 <- read_table2(
  "~/Documents/00_data_analysis/datasets/meteo_tandil/tandil.dat", 
  col_names = FALSE) %>% 
  rename(year=X1, julian=X2, tmax=X3, tmin=X4, rain=X5, hr=X6)   

serie_tandil <- transform(serie_tandil0, 
                          date = as.Date(paste0(year, "-1-1")) + julian - 1)   
serie_tandil <- serie_tandil%>% 
  mutate(month=lubridate::month(as.Date(date))) %>% 
  select(year, month, julian, date, everything())

serie_tandil %>% glimpse

tandil_2020_21 <- readxl::read_excel("~/Documents/00_data_analysis/datasets/meteo_tandil/TANDIL 2020 21.xlsx") %>% rename(date=Fecha, tmax="T.Max.", tmin="T.Min.", rain=Pp, hr=Humedad)   %>%
  mutate(date=dmy(date),
         year=year(date),
         month=month(date),
         julian=yday(date))

tandil_2020_21 %>% write_sheet(ss=tan, sheet = "weather")

tan20 <- read_sheet(tan, sheet = "weather") 
tan20$date <- as.Date(tan20$date, format =  "%Y-%m-%d")
tan20 %>% glimpse

# tandil_2020_21 %>% 
#   filter(year == 2021)

tandil_serie_71_21 <- serie_tandil %>% 
  filter(!year == 2021) %>% 
  filter(!year == 2020) %>% 
  bind_rows(tan20)  

# tandil_serie_71_21 %>% distinct(year) %>% pull()
# tandil_serie_71_21 %>% count(year)

serie_full <- tandil_serie_71_21 %>%  
  # mutate_at(vars(date), as.Date(date)) %>%
  mutate(tmin=replace(tmin, tmin<(-10), NA)) %>% 
  mutate(tmin = zoo::na.approx(tmin)) %>% 
  mutate(tmax=replace(tmax, tmax>40, NA)) %>% 
  mutate(tmax=replace(tmax, tmax<0, NA)) %>% 
  mutate(tmax = zoo::na.approx(tmax)) %>% 
  mutate(rain=replace(rain, rain<0, 0)) %>% 
  mutate(hr=replace(hr, hr<30, NA)) %>% 
  mutate(hr = zoo::na.approx(hr)) %>% 
  arrange(date)  

serie_full %>% glimpse
save(serie_full, file ="data/serie_tandil.RData" )
load("data/serie_tandil.RData" )

# serie_full%>%
# ggplot()+
#   aes(date, hr)+
#   geom_line()+
#   # geom_col()+
#   facet_wrap(~year, scales = "free")+
#   theme_void()+
#   geom_hline(yintercept = 0)
  

# serie_full %>% write_csv("data/tandil_serie71-20.csv")

# Serie climatica tandil 
# dat <- read.csv("https://raw.githubusercontent.com/juanchiem/agro_data/master/tandil_serie71-20.csv", sep = ",")
# dat$date <- as.Date(dat$date, format =  "%Y-%m-%d")
# dat %>% glimpse


```
