==> 03-manejo.Rmd <==
---
output:
  html_document:
    code_folding: hide
    chunk_output_type: console
---

# Manejo 

```{r message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE, message=FALSE, warning=FALSE)
pacman::p_load(tidyverse, lubridate, scales, ggridges)
# drive_auth(email = "edwardsmolina@gmail.com")
# gs4_auth(email = "edwardsmolina@gmail.com")
source(here::here("0-themes.R"))
load(here::here("data/data.Rdata"))
# glimpse(dat)
```



# Fechas de siembra por campaña - soja 1º

```{r}
soja_raw <-
  soja_raw %>% janitor::clean_names()
```

```{r}

#cow <- soja_raw %>% 
  #janitor::clean_names() %>%
  #mutate (fecha_de_siembra_dd_mm = lubridate::month(fecha_de_siembra_dd_mm)) %>%
  #mutate (echa_de_siembra_dd_mm = as.Date(fecha_de_siembra_dd_mm
  #                      , format = "%m.%d")
  #        , zona = as.factor(zona)
  #        ) %>%
  
  filter(#fecha_siembra > '2018-09-15', 
         #fecha_siembra < '2018-12-31', 
         soja_1_2 == "Soja 1°") %>%
  ggplot(aes(x = FDS, y = zona)) +
  geom_density_ridges(aes( stat ="density"
                           ,fill = campana
                           ), 
                      alpha =0.5,
                      jittered_points = TRUE, scale = .7, rel_min_height = .01,
                      point_shape = "|", point_size = 1, size = 0.25,
                      position = position_points_jitter(height = 0))+
  scale_x_date(date_labels = "%^b", breaks = date_breaks("1 months"), expand = c(0,0))+
  labs(x= "", y = "Zona", 
       title = "Evolución inter-anual de\nventanas de siembra - Soja 1°" ,
       fill = "")+
  theme_dens
```


```{r}
list.files(path = here::here("plots_manejo"), 
           pattern = ".png",
           all.files = TRUE, full.names = TRUE)%>% 
  map(~ drive_upload(.,
                     path = as_dribble("juanchi_guille/tandilia_fina_2020/manejo"),
                     overwrite = TRUE)
  )
```

```{r}
skimr::skim(dat)
```


Ventanas de siembra

```{r }
dat_fs <- dat %>% 
  drop_na(rinde) %>% 
  drop_na(zona_0) %>%      
  # filter(fecha_de_siembra_dia_mes > '2015-01-01') %>%  
  mutate(
    year = factor(lubridate::year(as.Date(fecha_de_siembra_dia_mes))),     
    date = update(as.Date(fecha_de_siembra_dia_mes), year = 1))
```

```{r }
dat_fs %>% 
  filter(!zona_0 == "Azul") %>% #view
  filter(cultivo_de_cosecha == "TRIGO") %>% 
  ggplot(aes(x = date, y = zona_0)) + 
  guides(fill=FALSE) +
  geom_density_ridges(aes(fill = zona_0),
                      panel_scaling = F, alpha =0.5,
                      jittered_points = TRUE, scale = .5, rel_min_height = .01,
                      point_shape = "|", point_size = 2, size = 0.25,
                      position = position_points_jitter(height = 0))+
  guides(fill = "none")+
  theme(axis.text.x = element_text(angle = 60, hjust = 1))+ 
  # facet_wrap("Zona")+
  labs(x= "", y = "",  
       title = "" )+  
  theme_dens
```

```{r eval = F}
ggsave(last_plot(), file = "plots_manejo/ventanas_siembra_trigo.png", width = 6, height = 5)
```

```{r }
dat_fs %>% #glimpse
  filter(cultivo_de_cosecha == "CEBADA") %>% 
  filter(date>"1-04-13") %>%
  # select(date)
  # filter(!zona_0 == "Azul") %>% #view
  ggplot(aes(x = date, y = zona_0)) + 
  guides(fill=FALSE) +
  geom_density_ridges(aes(fill = zona_0),
                      panel_scaling = F, alpha =0.5,
                      jittered_points = TRUE, scale = .5, rel_min_height = .01,
                      point_shape = "|", point_size = 2, size = 0.25,
                      position = position_points_jitter(height = 0))+
  guides(fill = "none")+
  theme(axis.text.x = element_text(angle = 60, hjust = 1))+ 
  # facet_wrap("Zona")+
  labs(x= "", y = "",  
       title = "" )+  
  theme_dens
```

```{r eval = F}
ggsave(last_plot(), file = "plots_manejo/ventanas_siembra_ceb.png", width = 6, height = 5)
```

```{r }
# dat %>% 
#   drop_na(rinde) %>% 
#   group_by(campana, Zona, cultivo_de_cosecha) %>% 
#   summarise(n_lotes = n(), 
#             rinde_med = quantile(rinde, 0.5), 
#             techo = quantile(rinde, 0.95)) -> lotes
# # unique(dat$variedad)
# dat %>%
#   filter(str_detect(cultivo_de_cosecha, 'Tr|Ceb')) %>%
#   count(cultivo_de_cosecha, variedad) %>%
#   arrange(cultivo_de_cosecha, desc(n)) #%>% 
# #  write_sheet(ss=jmf, sheet = "variedades")
```

```{r}
dat_fs %>% 
  filter(cultivo_de_cosecha == "TRIGO") %>% 
  filter(rinde > 4000) %>% 
  filter(!(zona_0== "Chillar-Barker" & rinde > 6500)) %>% 
  filter(!zona_0 == "Azul") %>% #view 
  # filter(between(rinde, 1000, 10000)) %>% 
  ggplot() +
  aes(date, rinde, color = zona_0, )+
  facet_wrap("zona_0")+
  geom_point(alpha = 0.2)+
  # geom_point(aes(color = year))+
  theme_dens1+
  geom_smooth(se = F, 
              method = 'lm',
              formula = y ~ poly(x, 1)
              ) +
  labs(x = "Fecha de siembra", y = "Rendimiento (kg/ha)", col ="", 
       title = "Rinde Trigo 2020")+
  guides(col = FALSE)
```

```{r eval = F}
ggsave(last_plot(), file = "plots_manejo/rinde_fs_trigo.png", width = 6, height = 5)

```

```{r}
options(scipen=999)
library(broom)

fits <- dat_fs %>% 
  filter(cultivo_de_cosecha == "TRIGO") %>% 
  filter(rinde > 4000) %>% 
  filter(!(zona_0== "Chillar-Barker" & rinde > 6500)) %>% 
  filter(!zona_0 == "Azul") %>% #view 
  select(zona_0, date, rinde) %>% 
  mutate_at(c("zona_0"), as.factor) %>%
  mutate(julian = yday(date)) %>% 
  nest(data = -zona_0) %>% 
  mutate(
    fit = map(data, ~ anova(lm(rinde ~ julian + I(julian^2), data = .x))),
    tidied = map(fit, tidy)
  )  

fits %>% 
  unnest(tidied) %>% 
  select(-data, -fit) %>% 
  knitr::kable()

fits %>% 
  mutate(tidy_model = map(fit, broom::tidy)) %>% #,
  # model_qual = map(model, MuMIn::r.squaredGLMM)) %>% 
  select(zona_0, tidy_model) %>%
  unnest(c(tidy_model)) 
out_fit
```


```{r}
dat_fs %>% #glimpsesummary(date)
  filter(!zona_0 == "Azul") %>% #view 
  filter(cultivo_de_cosecha == "CEBADA") %>% 
  filter(date>"1-04-13") %>%
  filter(rinde > 1000) %>% 
  mutate(zona_1 = case_when(zona_0 == "Chillar-Barker" ~ "Chillar-Barker", 
                            TRUE ~ "Serrana-Ayacucho")) %>% 
  # filter(!(zona_0== "Chillar-Barker" & rinde > 6500)) %>% 
  # filter(!zona_0 == "Azul") %>% #view 
  # filter(between(rinde, 1000, 10000)) %>% 
    ggplot() +
  aes(date, rinde)+  
  facet_wrap(~zona_1)+
  geom_point(aes(color=zona_0), alpha = 0.5)+
  geom_smooth(se = F, method = 'lm', formula = y ~ poly(x, 1)) +
  # geom_smooth(data = filter(dat_fs, Zona!="Costa"),
  #             se = F, method = 'lm', formula = y ~ poly(x, 1)) +
  labs(x = "Fecha de siembra", y = "Rendimiento (kg/ha)",
       col ="",
       title = "Rinde Cebada 2017-2020")+ 
  theme_dens1
```

```{r eval = F}
ggsave(last_plot(), file = "plots_manejo/rinde_fs_ceb.png", width = 6, height = 5)
```

```{r}
dat %>%  
  drop_na(rinde) %>% 
    filter(!zona_0 == "Azul") %>% #view 
  # filter(date > as.Date("0001-05-01") & 
  #        date < as.Date("0001-08-15")) %>% 
filter(between(rinde, 1000, 10000)) %>%   # mutate(zona = fct_rev(zona)) %>% 
  mutate(month = month(fecha_de_siembra_dia_mes, label = TRUE),
         day = day(fecha_de_siembra_dia_mes), 
         quincena = case_when(day %in% 1:15 ~ "Q1",
                              TRUE ~ "Q2"), 
         mes_q = factor(paste0(month, "-", quincena))) %>% 
  mutate(mes_q = reorder(mes_q, as.integer(month), .fun = identity)) %>% 
  droplevels() -> dat_quin
# dat_quin
```

```{r}
dat_quin %>% 
  drop_na(zona_0) %>% 
  filter(between(rinde, 3900, 10000)) %>%   # mutate(zona = fct_rev(zona)) %>% 
  filter(cultivo_de_cosecha == "TRIGO") %>% 
    ggplot(aes(x=mes_q, y=rinde))+
  # geom_boxplot(fill = "steelblue", outlier.size = 0, alpha=0.5) +
  geom_smooth(se = F, span = 0.8, col = "red3", linetype = 2, aes(group=1))+  
  facet_grid(~zona_0)+
  geom_point(pch = 21, size=0.5, aes(group=1))+
  labs(x = "Mes - Quincena", y = "Rendimiento (kg/ha)", 
       title = "TRIGO") + 
  theme_dens1
```

```{r eval = F}
ggsave(last_plot(),
       file = "plots/rinde_fs_quincena_tr.png", w =6, h=5)
drive_upload(path = as_dribble("juanchi_guille/JMF_fina_2020/plots"), "plots/rinde_fs_quincena_tr.png", overwrite = TRUE)
```

```{r}
dat_quin %>% 
  filter(!mes_q=="feb-Q1") %>%
  filter(cultivo_de_cosecha == "CEBADA") %>%
    mutate(zona_1 = case_when(zona_0 == "Chillar-Barker" ~ "Chillar-Barker", 
                            TRUE ~ "Serrana-Ayacucho")) %>% 
  ggplot(aes(x=mes_q, y=rinde))+
  geom_smooth(se = F, span = 0.8, col = "red3", linetype = 2, aes(group=1))+  
  facet_grid(zona_1~.)+
  geom_point(pch = 21, size=0.5)+
  labs(x = "Mes - Quincena", y = "Rendimiento (kg/ha)", 
       title = "Cebada") + 
  theme_dens1
```

```{r eval = F}
ggsave(last_plot(), 
       file = "plots_manejo/rinde_fs_quincena_ceb.png", w =6, h=5)
```

```{r}
(dat %>% 
  filter(variedad %in% 
           c("BAGUETTE_802", "BASILIO","BAGUETTE_501")) %>% 
  drop_na(fecha_de_siembra_dia_mes) %>%
  select(fecha_de_siembra_dia_mes, variedad,rinde, campana) %>% 
  group_by(campana) %>% #slice_sample(n = 3) %>%  
  mutate(
    year = factor(year(fecha_de_siembra_dia_mes)),     
    date = update(fecha_de_siembra_dia_mes, year = 1), 
    variedad = fct_relevel(variedad, "BAGUETTE_802", "BASILIO","BAGUETTE_501")) %>% 
  filter(!(date > '1-07-15' & variedad == "BAGUETTE_802")) %>% 
  filter(rinde < 10000) %>% 
  ggplot(aes(date, rinde)) +
  facet_grid(. ~variedad , scales="free_x")+
  geom_point()+
  # geom_point(aes(color = year))+
  theme_dens1+
  geom_smooth(se = F, method = 'lm', formula = y ~ poly(x, 2)) +
  labs(x = "Fecha de siembra", y = "Rendimiento (kg/ha)", col ="", 
       title = "Rinde trigo - ciclos (2017-2020)")
)
```

```{r eval = F}
ggsave(last_plot(), file = "plots/ciclos_trigo.png", w=5, h=5)
drive_upload(path = as_dribble("juanchi_guille/JMF_fina_2020/plots"), "plots/ciclos_trigo.png", overwrite = TRUE)
```

```{r}
dat %>% 
  filter(str_detect(cultivo_de_cosecha, 'TRIGO|CEB')) %>%
  filter(!str_detect(cultivo_de_cosecha, 'Candeal')) %>%
  # filter(!antecesor_estival %in% c("OTRO", NA)) %>%
  # filter(!str_detect(antecesor_estival, "SORGO|MAIZ_SEGUNDA|OTRO")) %>%
    filter(rinde>1000) %>%
  drop_na(antecesor_estival) %>% 
  ggplot(aes(antecesor_estival, rinde)) +
  facet_grid(.~cultivo_de_cosecha, scales="free_x")+
  geom_jitter(alpha=0.5, width = 0.1)+
  stat_summary(fun.data = "mean_cl_boot", colour = "red", size = 0.5)+  
  stat_summary(aes(label=round(..y..,0)), 
                     fun=mean, geom="text", size=4,vjust = -0.5, angle = 90)+

  geom_smooth(se = F, method = 'lm', formula = y ~ poly(x, 2)) +
  labs(x = "Fecha de siembra", y = "Rendimiento (kg/ha)", col ="", 
       title = "Rinde x antecesor (2017-2020)")+
  theme_dens1
```

```{r eval = F}
ggsave(last_plot(), file = "plots_manejo/antecesores.png", w=5, h=5)

```

# Cultivo de servicio 

```{r}
cs
```

```{r}
cs %>% 
  # group_by(campana) %>%
  # mutate(row = row_number()) %>%
  ggplot() +
  aes(x=rto_barbecho, y=rto_cultivo_de_servicio-rto_barbecho)+
  facet_grid(~ campana, scales = "free")+
  # facet_grid(~ campana, ncol=4, scales = "free")+
  geom_point(alpha=.3)+
  # geom_smooth(se = F, method = 'lm', formula = y ~ poly(x, 1)) +
  geom_smooth(se = F, span=.8, col ="black")+
  geom_hline(yintercept=0)+
  labs(x= "Rinde barbecho (qq/ha)", y = "Dif rinde CS - Barb (qq/ha)")+
  ylim(-20,20)+
  xlim(10,50)+
  geom_rect(aes(xmin = -Inf, xmax = Inf, ymin = -Inf, ymax = 0),
            alpha = .002, fill = "blue") +
  geom_rect(aes(xmin = -Inf, xmax = Inf, ymin = 0, ymax = Inf),
            alpha = .002, fill = "red")+ 
  expand_limits(x = 0, y = 0)+
  theme_minimal()
  
ggsave(last_plot(), file = "03-manejo/cs_bar_campana.png", w=7, h=2.5)
```
