==> 04-genetica.Rmd <==
# GM soja - Pie chart
```{r}
raw_sum2 <- raw %>% 
  group_by(zona, soja_1_2, gm) %>%
  summarise(#var_lotes= n(), 
    var_sup = sum(superficie_ha, na.rm = TRUE),
    rinde_medio = median(rendimiento_qq_ha, na.rm = TRUE), 
    .groups = 'drop') %>% 
  mutate_if(is.character, as.factor) %>% 
  drop_na() %>% 
  group_by(zona, soja_1_2) %>% 
  mutate(porcent=var_sup/sum(var_sup)*100)

raw_sum2 %>% #data.frame
  ggplot(aes(x =factor(1), y = porcent, fill = gm)) +
  geom_bar(stat='identity') +
  coord_polar(theta = "y")+
  facet_grid(soja_1_2 ~ zona, switch = "y")+
  theme_void()+
  geom_text(aes( label=paste0(round(porcent), "\n",gm)), 
            position=position_stack(vjust=0.5), color="white",size=3, fontface ="bold")+
  theme(legend.position="bottom")+
  scale_fill_viridis_d(direction = -1)

ggsave(last_plot(), file = "plots/gm_zon.png", width = 7, height = 4)
```


# Variedades

```{r message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE, message=FALSE, warning=FALSE)
pacman::p_load(tidyverse)
# pacman::p_load(tidyverse, googlesheets4, googledrive)
# javi <- gs4_get("1jTFU8fQw2fiTyfrhuPSjc4lcAC6U9G-zjvOlADQ6JNE")
# gs4_browse(javi)
# drive_auth(email = "edwardsmolina@gmail.com")
# gs4_auth(email = "edwardsmolina@gmail.com")
source(here::here("0 themes.R"))
load(here::here("data/data.Rdata"))
# glimpse(dat)
```

```{r eval=F}
list.files(path = here::here("plots_variedades"), 
           pattern = ".png",
           all.files = TRUE, full.names = TRUE)%>% 
  map(~ drive_upload(.,
                     path = as_dribble("juanchi_guille/tandilia_fina_2020/variedades"),
                     overwrite = TRUE)
  )
```

```{r}
# dat %>%
#   filter(str_detect(cultivo_de_cosecha, 'Tr|Ceb')) %>%
#   select(cultivo_de_cosecha, campana,zona_0, variedad, rinde, superficie ) %>% 
#   group_by(cultivo_de_cosecha, zona_0, variedad) %>%
#   # filter(n() > 5) %>% ungroup %>% 
 #   mutate_if(is.character, as.factor) %>% 
#   group_by(cultivo_de_cosecha, variedad) %>%
#   summarise(#var_lotes= n(), 
#     var_sup = sum(superficie, na.rm = TRUE),
#     rinde_medio = median(rinde, na.rm = TRUE), 
#     .groups = 'drop') %>% 
#   group_by(cultivo_de_cosecha) %>% 
#   mutate(porcent=var_sup/sum(var_sup)*100) %>% ungroup %>% 
#   arrange(cultivo_de_cosecha, -porcent)-> dat_var
# dat_var
```

```{r}
# dat_var %>% #data.frame
#   filter(cultivo_de_cosecha =="Cebada") %>% 
#   ggplot(aes(x =factor(1), y = porcent, fill = variedad)) +
#   geom_bar(stat='identity') +
#   coord_polar(theta = "y")+
#   facet_grid(~zona_0, switch = "y")+
#   theme_void()+
#   geom_text(aes(label=paste0(round(porcent))), 
#             position=position_stack(vjust=0.5), color="white",size=3, fontface ="bold")+
#   theme(legend.position="bottom")+
#   scale_fill_viridis_d(begin = 0, end =0.8, direction = -1)

# ggsave(last_plot(), file = "plots/sup_var_zona_0_ceb.png", width = 7, height = 4)
# drive_upload(path = as_dribble("juanchi_guille/JMF_fina_2020"), "plots/sup_var_zona_0_ceb.png", overwrite = TRUE)

```

```{r eval =F}
# ggsave(last_plot(), 
#        file = "plots_variedades/sup_cebada.png", 
#        width = 7, height = 4)
# drive_upload(
#   "plots_variedades/sup_cebada.png", 
#   path = as_dribble("juanchi_guille/tandilia_fina_2020/superficies"),
#   overwrite = TRUE)
```


```{r}
# dat_var %>% #data.frame
#   filter(cultivo_de_cosecha =="Trigo") %>% 
#   drop_na() %>% 
#   ggplot(aes(x =factor(1), y = porcent, fill = variedad)) +
#   geom_bar(stat='identity') +
#   coord_polar(theta = "y")+
#   facet_grid(~zona_0, switch = "y")+
#   theme_void()+
#   geom_text(aes(label=paste0(round(porcent))), 
#             position=position_stack(vjust=0.5), color="white",size=3, fontface ="bold")+
#   theme(legend.position="bottom")+
#   scale_fill_hue()

# ggsave(last_plot(), file = "plots/sup_var_zona_0_tr1.png", width = 7, height = 4)
```

```{r}
# dat_var %>% 
#   filter(cultivo_de_cosecha == "Trigo") %>%
#     # filter(variedad == "Andreia") 
#   # drop_na(rinde) %>%
#   ggplot(aes(x = factor(campana), 
#              y = porcent, 
#              col = variedad, 
#              group =variedad))+
#   geom_line()+geom_point()+
#   facet_grid(zona_0~.)+
#   labs(x = NULL, y = "% del área de cada zona_0",  col="",
#        title = "Evolucion de superficie de trigo")+
#   # guides(col="none")+
#   ggrepel::geom_text_repel(
#     data = dat_var %>% 
#         filter(cultivo_de_cosecha == "Trigo") %>%
#       filter(campana == "19-20"),
#       # drop_na(tecno_sts),
#     aes(label = variedad), size =3, nudge_x = 0.1) +
#   theme_dens1
  
# ggsave(last_plot(), file = "plots/evolucion_trigo.png", w =7, h=4)
```

## Superficies (tamaño de cuadrados) y rendimientos (intensidad de color) 

```{r}
library(treemap)
```

```{r }
dat_sum <- dat %>% 
  # filter(cultivo_de_cosecha == "Trigo") %>% 
  # filter(campana == "19-20") %>% 
  # group_by(variedad) %>%
  # filter(n()>10) %>% ungroup %>%
  group_by(cultivo_de_cosecha, variedad) %>%
  summarise(lotes= n(), 
            sup = sum(superficie, na.rm = TRUE),
            rinde_medio = median(rinde, na.rm = TRUE),
            .groups = 'drop') %>% 
   group_by(cultivo_de_cosecha) %>% 
   mutate(porcent=sup/sum(sup)*100) %>% ungroup %>% 
  mutate_if(is.character, as.factor)%>%
  mutate(label = paste(variedad, round(sup,0), sep = "\n"))
dat_sum
```


```{r }
dat_sum %>%
    write_sheet(ss=tan, sheet = "superficies")
```


```{r }
```

```{r }
png(filename="plots_variedades/treemap_trigo.png",
    width = 120, height = 120, units='mm', res = 300)

subset(dat_sum, 
       cultivo_de_cosecha == "TRIGO"  
       # campana == '17-18'
       ) %>% 
  treemap(index="label",
          vSize="sup",
          vColor="rinde_medio",
          # border.col=c("grey70", "grey90"),
          # palette="Spectral",
          type = "value",
          title = "Superficie de trigo (has) - 2020",
          title.legend = "Rendimiento medio (qq/ha)",
          overlap.labels=1, 
          bg.labels=c("transparent"),              
          inflate.labels=F,
          align.labels=c("center", "center")) 
dev.off()

# drive_upload(path = as_dribble("juanchi_guille/JMF_fina_2020"), "plots/treemap_trigo_17-18.png", overwrite = TRUE)
```


```{r}
png(filename="plots_variedades/treemap_cebada.png",
    width = 120, height = 120, units='mm', res = 300)

dat_sum %>% 
  filter(cultivo_de_cosecha == "CEBADA") %>%  
  filter(!variedad == "YARARA") %>% 
treemap(index=c("label"),
          vSize="sup",
          vColor="rinde_medio",
          # border.col=c("grey70", "grey90"),
          # palette="Spectral",
          type = "value",
          title = "Superficie de cebada 2020 (has)",
          title.legend = "Rendimiento medio (qq/ha)",
          overlap.labels=1, 
          bg.labels=c("transparent"),              # Background color of labels
          inflate.labels=F,
           align.labels=list(
        c("left", "top"),
        c("center", "center"))) 
dev.off()
# 
# drive_upload(path = as_dribble("juanchi_guille/JMF_fina_2020"), "plots/treemap_cebada.png", overwrite = TRUE)

```

## Performance variedades

```{r}
dat %>%
  filter(str_detect(cultivo_de_cosecha, 'TRIGO|CEB')) %>%
  select(cultivo_de_cosecha, variedad,rinde ) %>% 
  group_by(cultivo_de_cosecha, variedad) %>%
  filter(n() > 3) %>% ungroup %>% 
  mutate_if(is.character, as.factor) ->dat1

dat1 %>% 
  filter(str_detect(cultivo_de_cosecha, 'TRIGO')) %>%
  # filter(campana == "20-21") %>% 
  distinct(variedad) %>% 
  pull() %>% 
  droplevels()-> var_20_tr

dat1 %>% 
  group_by(cultivo_de_cosecha, variedad) %>% 
  summarise(n=n(), 
            rinde_medio = mean(rinde)) %>% 
  arrange(cultivo_de_cosecha, -n) %>% 
  write_sheet(ss=tan, sheet = "var_rindes")
```

```{r}
dat1 %>% 
filter(cultivo_de_cosecha== "TRIGO" &
                 variedad %in% var_20_tr) %>%   
        ggplot()+
        aes(x=reorder(variedad, rinde, mean), y=rinde)+
        geom_point(alpha=0.3)+
        stat_summary(fun.data = "mean_cl_boot", colour = "red", size = 0.5)+
        coord_flip()+
        theme_bw2+
        stat_summary(aes(label=round(..y..,0)), 
                     fun=mean, geom="text", size=4,vjust = -0.5)+
        scale_y_continuous(breaks= scales::pretty_breaks())+
        labs(title = "Performance variedades trigo", 
             x="", y="kg/ha")
```

```{r}
last_plot() %>%  
  ggsave(file = "plots_variedades/trigo_perf.png", w=5, h=4)
```

```{r}
dat1 %>% 
filter(cultivo_de_cosecha== "CEBADA") %>%   
        ggplot()+
        aes(x=reorder(variedad, rinde, mean), y=rinde)+
        geom_point(alpha=0.3)+
        stat_summary(fun.data = "mean_cl_boot", colour = "red", size = 0.5)+
        coord_flip()+
        theme_bw2+
        stat_summary(aes(label=round(..y..,0)), 
                     fun=mean, geom="text", size=4,vjust = -0.5)+
        scale_y_continuous(breaks= scales::pretty_breaks())+
        labs(title = "Performance variedades cebada", 
             x="", y="kg/ha")
```

```{r, eval=FALSE}
last_plot() %>%  
  ggsave(file = "plots_variedades/cebada_perf.png", w=5, h=4)
```
