==> 04-geno_table.Rmd <==
---
title: "Tabla genotipos"
output: html_document
---

```{r message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE, message=FALSE, warning=FALSE)
pacman::p_load(tidyverse, formattable)
# pacman::p_load(tidyverse, googlesheets4, googledrive)
# javi <- gs4_get("1jTFU8fQw2fiTyfrhuPSjc4lcAC6U9G-zjvOlADQ6JNE")
# gs4_browse(javi)
# drive_auth(email = "edwardsmolina@gmail.com")
# gs4_auth(email = "edwardsmolina@gmail.com")
```


```{r message=FALSE, warning=FALSE}
source(here::here("0-themes.R"))
source(here::here("geno_new.R"))
load(here::here("data/data.Rdata"))
```

```{r cars}
glimpse(dat)
# https://stackoverflow.com/questions/50258443/fast-way-for-string-matching-and-replacement-from-another-dataframe-in-r
```

```{r}
dat_geno %>% 
  count(campana, gm, geno)%>% 
  group_by(campana, gm) %>% 
  mutate(porcent=n/sum(n)*100) %>%  
  # dplyr::filter(campana =="2020/21")
  select(-n) %>%
  # arrange(campana, gm, geno) %>%
  pivot_wider(names_from = campana,
              values_from = porcent) %>%
  # mutate(across(where(is.numeric), round, 2)) %>% 
  # mutate(change = (`2020/21` - `2019/20`) %>% round(1)) %>% 
  arrange(gm, -`2020/21`) %>% 
  mutate_if(is.numeric, replace_na, 0) %>% 
  mutate(across(where(is.numeric), round, 1)) ->dat_geno_table
```

GM III

```{r}
dat_geno_table %>% 
  ungroup() %>% 
  mutate(tot = rowSums(dplyr::select_if(., is.numeric), na.rm = TRUE)) %>%
  filter(tot>5) %>%   
  filter(gm=="III") %>% 
  formattable(align = "r",
              list(geno=
                     formatter("span", style = ~ style(color = "grey",font.weight = "bold")),
                   area(col = 3:6) ~ function(x) percent(x / 100, digits = 1),
                   area(col = 3:6) ~ color_bar("#B1CBEB")
                   # change = 
                   #   formatter(
                   #     "span",
                   #     style = x ~ style(color = case_when(
                   #       x < 0 ~ "red", 
                   #       x > 0 ~ "green",
                   #       TRUE ~  "blue"
                   #     )),
                   #     x ~ icontext(case_when(
                   #       x < 0 ~ "arrow-down", 
                   #       x > 0 ~ "arrow-up",
                   #       TRUE ~  ""), x))
              )
  )
```

GM IV

```{r}
dat_geno_table %>% 
  ungroup() %>% 
  mutate(tot = rowSums(dplyr::select_if(., is.numeric), na.rm = TRUE)) %>%
  filter(tot>5) %>% 
  filter(gm=="IV") %>% 
  formattable(align = "r",
              list(geno=
                     formatter("span", style = ~ style(color = "grey",font.weight = "bold")),
                   area(col = 3:6) ~ function(x) percent(x / 100, digits = 1),
                   area(col = 3:6) ~ color_bar("#B1CBEB")
              )
  )
```

GM V

```{r}
dat_geno_table %>% 
  ungroup() %>% 
  mutate(tot = rowSums(dplyr::select_if(., is.numeric), na.rm = TRUE)) %>%
  filter(tot>5) %>% 
  filter(gm=="V") %>% 
  formattable(align = "r",
              list(geno=
                     formatter("span", style = ~ style(color = "grey",font.weight = "bold")),
                   area(col = 3:6) ~ function(x) percent(x / 100, digits = 1),
                   area(col = 3:6) ~ color_bar("#B1CBEB")
              )
  )
```

GM VI

```{r}
dat_geno_table %>% 
  ungroup() %>% 
  mutate(tot = rowSums(dplyr::select_if(., is.numeric), na.rm = TRUE)) %>%
  filter(tot>5) %>% 
  filter(gm=="VI") %>% 
  formattable(align = "r",
              list(geno=
                     formatter("span", style = ~ style(color = "grey",font.weight = "bold")),
                   area(col = 3:6) ~ function(x) percent(x / 100, digits = 1),
                   area(col = 3:6) ~ color_bar("#B1CBEB")
              )
  )
```

GM VII

```{r}
dat_geno_table %>% 
  ungroup() %>% 
  mutate(tot = rowSums(dplyr::select_if(., is.numeric), na.rm = TRUE)) %>%
  filter(tot>5) %>%   filter(gm=="VII") %>% 
  formattable(align = "r",
              list(geno=
                     formatter("span", style = ~ style(color = "grey",font.weight = "bold")),
                   area(col = 3:6) ~ function(x) percent(x / 100, digits = 1),
                   area(col = 3:6) ~ color_bar("#B1CBEB")
              )
  )
```


Note that the `echo = FALSE` parameter was added to the code chunk to prevent printing of the R code that generated the plot.
