==> 06-malezas.Rmd <==
# Malezas

```{r message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE, message=FALSE, warning=FALSE)
pacman::p_load(tidyverse, fuzzyjoin) 
load(here::here("data/data.Rdata"))
source(here::here("0-themes.R"))

```

```{r}
dat %>% 
  filter(soja_1_2 == "SOJA_1°") %>%  
  select(campana, zona, especie_principal, especie_secundaria) %>% 
  mutate(across(starts_with("especie"), 
                ~stringi::stri_trans_general(., "Latin-ASCII"))) %>% #sacar acentos
  mutate(across(starts_with("especie"), 
                ~str_to_upper(gsub(',', '\\.',.)))) %>%  # UPPER + sacar ","
  mutate(across(starts_with("especie"), 
                ~str_replace(., "SP.", ""))) %>%  # sacar SP.
  mutate(across(starts_with("especie"), 
                ~str_trim(.))) -> adv_herb   #sacar espacios antes-desp

adv_herb
```

```{r}
adv_herb %>% 
  drop_na(especie_principal) %>% 
  count(zona, campana, especie_principal) %>% 
  ungroup() %>% 
  group_by(zona, campana) %>%
  arrange(zona, campana, -n) %>% 
  top_n(3, n) %>% 
  mutate(especie_principal=str_replace(especie_principal, "_", " ")) %>% 
  mutate(especie_principal=str_replace(especie_principal, "_", " "))-> weeds1
weeds1
```

```{r}
adv_herb %>% 
  drop_na(especie_secundaria) %>% 
  count(zona, campana, especie_secundaria) %>% 
  ungroup() %>% 
  group_by(zona, campana) %>%
  arrange(zona, campana, -n) %>% 
  top_n(3, n) %>% 
  mutate(especie_secundaria=str_replace(especie_secundaria, "_", " ")) %>% 
  mutate(especie_secundaria=str_replace(especie_secundaria, "_", " "))-> weeds2
weeds2
```


```{r}
adv_herb %>%
  filter_at(vars(starts_with("especie")), any_vars(!is.na(.))) %>%
  pivot_longer(cols= -(campana:superficie_ha), 
               names_to = "nivel", 
               values_to = "maleza")%>% 
  count(maleza) 

malezas_ref <- gsheet::gsheet2tbl("https://docs.google.com/spreadsheets/d/1cTzo36jHRULfBf8c83TDpauxiwHrxeKrGprUUxLRzm0/edit#gid=1182890772")

malezas_ref %>% print(n=Inf)

# browseURL("https://docs.google.com/spreadsheets/d/1YL6vwgVu1nyMuZLRZw5uMav7cMt2mDXd7mI2S8Q6Q7c/edit#gid=950522321")

adv_herb %>% #count(especie_principal)
  filter(soja_1_2 == "SOJA_1°") %>%  
  filter_at(vars(starts_with("especie")), any_vars(!is.na(.))) %>%
  pivot_longer(-(campana:superficie_ha), names_to = "nivel", 
               values_to = "maleza")%>% 
  filter(!maleza %in% c(NA, "OTRO", "OTRA", "VARIAS", "ROSETA")) %>% 
  stringdist_left_join(malezas_ref %>% drop_na(genero), 
                       by = c(maleza = "genero"), max_dist = 1) %>%
  stringdist_left_join(malezas_ref %>% drop_na(n_comun), 
                       by = c(maleza = "n_comun"), max_dist = 1) %>% view
  mutate(maleza = coalesce(n_comun.x,n_comun.y )) %>%
  select(campana:nivel, maleza)%>%
  count(campana, zona, nivel, maleza)%>% 
  # summarise(var_sup = sum(superficie_ha, na.rm = TRUE))%>%  
  group_by(campana, zona, nivel) %>% 
  mutate(zona= as.factor(zona),
         porcent=var_sup/sum(var_sup)*100, 
         nivel = recode(nivel, especie_principal = "1° Especie", 
                        especie_secundaria = "2° Especie", 
                        especie_terciaria= "3° Especie")) %>% 
  drop_na(maleza) -> herb_long
herb_long
```

```{r}
# herb_long %>% write_sheet(ss=sheet_2017_2021_cba_soja, sheet = "malezas")
herb_long<- googlesheets4::read_sheet(sheet_2017_2021_cba_soja, 
                  sheet = "malezas", guess_max = 10000, skip=0) %>% 
  janitor::clean_names() 
```

```{r}
plots_herb <- 
  herb_long %>%
  filter(!nivel == "3° Especie")%>%
  group_by(zona) %>%
  nest() %>%
  mutate(plot = map2(data, zona, 
                     ~ggplot(data = .x,  
         aes(x = factor(campana), 
             y = porcent, 
             col = maleza, 
             group = maleza))+
             geom_line()+
  geom_point()+
  labs(x = NULL, y = "% del área de cada zona",
        title = paste0("Evolución de principales malezas -  Zona ", .y))+
  guides(col="none")+
  facet_wrap(~nivel)+
  ggrepel::geom_text_repel(
    data = .x  %>% filter(zona == .y, campana == "2020/21"),
    aes(label = maleza), size =3, nudge_x = 0.2) +
   ggrepel::geom_text_repel(
    data = .x  %>% filter(zona == .y, campana == "2017/18"),
    aes(label = maleza), size =3, nudge_x = 0.2) +
   theme_bw2
))
plots_herb$plot
# file_names <- paste0("adversidades/", levels(herb_long$Zona), "_malezas.png")
# map2(file_names, plots_herb$plot, ggsave) 
```
